Change route to fix error
